requirements.txt
python-telegram-bot==20.7
notion-client==2.0.0
python-dotenv==1.0.0
pytest==7.4.3
asyncio==3.4.3